+++ src/org/opencms/ui/contextmenu/CmsSubmenu.java
@@ -49,7 +49,7 @@ public class CmsSubmenu implements I_CmsContextMenuItem {
     protected String m_parentId;
 
     /** The order. */
-    protected int m_order;
+    protected float m_order;
 
     /** The priority. */
     protected int m_priority;
@@ -66,7 +66,7 @@ public class CmsSubmenu implements I_CmsContextMenuItem {
      * @param order the order
      * @param priority the priority
      */
-    public CmsSubmenu(String id, String parentId, String title, int order, int priority) {
+    public CmsSubmenu(String id, String parentId, String title, float order, int priority) {
         m_id = id;
         m_parentId = parentId;
         m_title = title;
@@ -93,7 +93,7 @@ public class CmsSubmenu implements I_CmsContextMenuItem {
     /**
      * @see org.opencms.ui.contextmenu.I_CmsContextMenuItem#getOrder()
      */
-    public int getOrder() {
+    public float getOrder() {
 
         return m_order;
